default policy: Move some tests to integration
